requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf==3.17.4
python-docx==1.1.0
markdown==3.5.1
unstructured==0.11.8
google-generativeai==0.3.1 
python-dotenv==1.0.0         
tiktoken==0.5.2
streamlit==1.29.0
docx2txt==0.8
pandas==2.1.4
requests==2.31.0
numpy>=1.24.0          